Almost forgot to split up the components - I haven't rethought the interactions and breakdown between the components, I've just split the file up.
