+++ lib/components/sharedStyles.ts
@@ -1,0 +1,15 @@
+export const buttonClasses =
+  "bg-blue-500 hover:bg-blue-600 text-white font-semibold py-2 px-4 rounded-md transition-colors mr-6 mt-6 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-opacity-50";
+export const normalTextClasses = "text-gray-700 text-lg mb-4";
+export const headingClasses =
+  "text-3xl font-bold mb-8 text-gray-800 text-center pt-16";
+export const normalFormClasses =
+  "bg-white p-6 rounded-lg shadow-lg mb-16 border-gray-300";
+export const normalAreaClasses = normalFormClasses;
+export const normalLinkClasses =
+  "text-blue-500 hover:underline focus:outline-none focus:ring-2 focus:ring-blue-500";
+export const normalLabelClasses = "block text-gray-700 text-lg font-bold mb-2";
+export const normalInputClasses =
+  "w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500";
+export const roomUrlClasses =
+  `newlyCreatedRoomUrl mt-6 mb-8 ${normalTextClasses}`;
