+++ routes/static-routes.tsx
@@ -1,6 +1,7 @@
 import { RouteHandler, Routes } from "../lib/types.ts";
 import { playgroundWrapReactElem } from "./helpers.tsx";
-import { HomePage, NotFoundPage } from "../web-server/components.tsx";
+import { HomePage } from "../lib/components/HomePage.tsx";
+import { NotFoundPage } from "../lib/components/NotFoundPage.tsx";
 
 export const staticRoutes: Routes = {
   "/": {
